Move blob options to their own module
