+++ sdk/storage_blobs/src/blob/operations/update_page.rs
@@ -1,4 +1,4 @@
-use crate::{prelude::*, BA512Range};
+use crate::prelude::*;
 use azure_core::{headers::*, prelude::*, RequestId};
 use azure_storage::{headers::content_md5_from_headers, ConsistencyMD5};
 use bytes::Bytes;
