--- sdk/storage_blobs/src/headers.rs
@@ -1,3 +1,0 @@
-use azure_core::headers::HeaderName;
-
-pub const REHYDRATE_PRIORITY: HeaderName = HeaderName::from_static("x-ms-rehydrate-priority");
