+++ sdk/storage_blobs/src/options/blob_content_encoding.rs
