+++ sdk/storage_blobs/src/options/condition_max_size.rs
