+++ sdk/storage_blobs/src/options/delete_snapshot_method.rs
