+++ sdk/storage_blobs/src/options/hash.rs
