+++ sdk/storage_blobs/src/options/snapshot.rs
