+++ sdk/storage_blobs/src/prelude.rs
@@ -1,12 +1,10 @@
 pub use super::container::PublicAccess;
+pub use crate::options::*;
 pub use crate::{
     blob::{Blob, BlobBlockType, BlockList, BlockListType},
     clients::{
         AsBlobServiceClient, AsContainerClient, BlobClient, BlobLeaseClient, BlobServiceClient,
         ContainerClient, ContainerLeaseClient,
     },
-    AccessTier, BlobCacheControl, BlobContentDisposition, BlobContentEncoding, BlobContentLanguage,
-    BlobContentMD5, BlobContentType, BlobVersioning, BlockId, ConditionAppendPosition,
-    ConditionMaxSize, DeleteSnapshotsMethod, Hash, RehydratePriority, Snapshot, VersionId,
 };
 pub use azure_storage::core::{StoredAccessPolicy, StoredAccessPolicyList};
